Remove the BaseLivePayload class
* Merge the code for tar and image installations.
* Update the kernel version list right before the installation.
